bundle: update submodules
pathogen gets updated too.

This isn't quite a `git submodule foreach git pull` because I attempted
to individually test each one as it came in (which is not possible if I
just blindly pull them all).

On a related note, I'm not updating vim-ruby to HEAD.
This is because I accuse f253df2c44828927048f7daac4a8753cb472e689 of
having bad performance.
Will have to figure out how to deal with that. Don't want to have to
create a fork that reverts just that one commit.